+++ data/src/main/java/ddd/buyornot/data/model/login/KaKaoAuthRequest.kt
@@ -1,0 +1,5 @@
+package ddd.buyornot.data.model.login
+
+data class KaKaoAuthRequest(
+    val authorizationCode: String
+)
